+++ geozilla/geozilla-core/source/core/GeozillaCore.h
@@ -1,4 +1,4 @@
 #pragma once
 
-__declspec(dllexport) const char* GenerateGeoJson(const char* path, double latitude, double longitude);
+__declspec(dllexport) const char* GenerateGeoJson(const char* path, float latitude, float longitude);
 __declspec(dllexport) void FreeBuffer(const char* buffer);
